Update meshblock parameter types and enhance error handling in file reading functions. Changed meshblock index type from size_t to int for consistency and added checks to prevent out-of-bounds access. Improved resource management by ensuring file pointers are closed on error conditions.

+++ model/meshblocks/model_params.h
@@ -18,7 +18,7 @@ int read_line(FILE *fp, char **line, char *message);
 int read_int(FILE *fp, size_t bytes);
 double read_double(FILE *fp, size_t bytes);
 
-int is_within_meshblock(size_t mb, double x1, double x2, double x3);
+int is_within_meshblock(int mb, double x1, double x2, double x3);
 int get_meshblock(double x1, double x2, double x3);
 
 #endif // MODEL_PARAMS_H
